doc&refactor: missing changes from #248  (#249)
* feat: add GetAccountDetails endpoint to tonic rpc client

* feat: update CLI to provide account storage mode as cli args

* feat: handle updated onchain accounts during sync and refactor

Now during the state sync we check the updated account hashes for onchain
accoints. if they don't match, then we fetch the current state from the node.
The old validation for offchain accounts is left unchanged but I refactored it
so the logic looks similar to the one for onchain accoints. I also left a few
TODO's and panics that'll be removed in follow-up commits.

* use updated create_basic_wallet and faucet

* use next for miden node proto

* update order of create_basic_fungible_faucet

* add small integration test for onchain accounts transactions

* address comments

* add missing sync

* fix: read Account instead of AccountDetails from get_account_details response

* remove dbg statement

* re-enable other integration tests

* extend mint integration test to reuse same on-chain account in two clients

* Test fix

* Test fix

* Test fix

* Update Makefile.toml

* address review

* extend onchain integration test to do transfer between accounts

* update missing documentation

* fix test

* Follow ups

* Stop test execution before transfer

We stop the onchain accounts test before running the p2id transaction because there's an error on how inputs are inserted/extracted on the vm, so until that gets fixed we'll leave the test code but do an early return so the test still passes.

* add rpc endpoints to trait functions' documentation

* split test helper in two functions

* fix tests

* Update main.rs

* dep version

* re-add assert

* remove unnecessary sleeps

---------

Co-authored-by: Ignacio Amigo <ignacio.amigo@lambdaclass.com>