refactor: use transaction IDs from sync response
After miden-node's changes in #377, we now have access to the transaction ids that were included in blocks since the last sync height. This enables us to remove our custom logic. In this commit we also leave the code ready for when the endpoint also returns the block height of each transaction, because now we don't have that info and assume it's always the response's block height. After #380 from miden-node gets address we'll start receiving that info in some way.