test: fix flakyness on checking tracked notes count
In some tests we did some assertions by doing `client.get_input_notes(filter).unwrap()` and checking the length of it. While in most cases that is enough, it will sporadically become an error if the node running the integration test has notes whose tag was built with another account's id that also has the same prefix (and hence the same tag as the ones created for the actually desired account). In order to mitigate it, we do a filter by account_id to ensure the note we want is there (there might be others, we won't care too much). In the case of swap tests, I also used `get_consumable_notes` since in that situation we wanted to consume the notes anyways.